requirements.txt
# ---------------------------
# Core Web Framework
# ---------------------------
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1

# ---------------------------
# LLM Provider (Groq)
# ---------------------------
groq==0.9.0

# ---------------------------
# LangChain (VERSION-ALIGNED)
# ---------------------------
langchain-core==0.2.27
langchain-community==0.2.11
langchain-text-splitters==0.2.2

# ---------------------------
# Embeddings & Reranking
# ---------------------------
sentence-transformers==2.7.0
torch>=2.1.0

# ---------------------------
# Vector Search
# ---------------------------
faiss-cpu==1.8.0.post1
numpy==1.26.4

# ---------------------------
# Hybrid Search (BM25)
# ---------------------------
rank-bm25==0.2.2

# ---------------------------
# Document Processing
# ---------------------------
pypdf==4.2.0

# ---------------------------
# Streaming / Async
# ---------------------------
anyio==4.4.0
httpx==0.27.0

# ---------------------------
# Evaluation
# ---------------------------
ragas==0.1.9
deepeval==0.21.30

# ---------------------------
# Logging / Token Tracking
# ---------------------------
structlog==24.1.0
tiktoken==0.7.0

# ---------------------------
# Utilities
# ---------------------------
rich==13.7.1
